+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Configure port mapping for the browser-sync server.
 - Configure shared folders for hosts configuration.
 - Configure custom shared folder mappings.
+- Configure the server's timezone.
 
 ### Changed
 - Rename `hosts` option to `webserver`.